+++ pom.xml
@@ -7,7 +7,7 @@
 	<version>0.0.1-SNAPSHOT</version>
 	<packaging>jar</packaging>
 
-	<name>WordCount</name>
+	<name>GenerateListOfOriginals</name>
 	<url>http://maven.apache.org</url>
 
 	<properties>
